Add a transform queue on the connector
If the target's transform queue is processing, then we should queue up the
transform on the connector instead of on the target.

This ensures that comparisons are made against the target's most up to
date state. Note that the connector's queue processing is triggered
by the `didComplete` event for the target's queue.