Management of inverse links moved to Source#_transform
This responsibility has been moved from `addLink` and `removeLink`
so that it can be centralized in `transform` and work regardless of the
entry point.

Furthermore, not all sources should apply inverse links directly. For
instance, remote sources such as JSONAPISource can rely on servers to
apply inverse links and then notify the source of the change.