Settle each transform completely before processing the next queued transform.
While a transform's `didTransform` events are processed, ensure that resulting
transforms and subsequent events are settled together. In this way, the full
effects of a transform will be known before the next queued transform is
applied.

This allows consecutive transforms to be applied to a Transformable source
without explicitly settling each transform.

This change also eliminates the need for a separate transformQueue to be
maintained on the TransformConnector.

[Closes #60]