Convert ActionQueue to work with Actions.
ActionQueue is now more flexible, and not just tied to a single
function to be called to process actions.

Instead, Action objects are processed sequentially.

The ill-advised `then` method has been removed in favor of returning
a promise from `process`.

Events are emitted when each action is finished processing 
(`actionComplete`) and when the queue is completely done 
(`queueComplete`).