Introduce Transformation.
Transformations are created implicitly by Transformable sources in order
to group together related operations (i.e. those with a shared history).
This allows related operations to be processed together, and unrelated 
operations to be queued in separate transformations to be processed
later.

A transformation tracks the original operations pushed to it, and can
verify whether other operations are descendants of those originals.