Introduce 3 memory source sync blocking test.
This test suite connects three memory sources with synchronous blocking
bi-directional transform connectors.

This configuration should not be used when more than one source can
spontaneously mutate, out of the context of responding to another source's
mutations. For example, socket sources receive data spontaneously. Having
more than one in this configuration could create a deadlock in a central
source that is receiving data from both and sending it on to the other.
In such a scenario, it would be preferable to use non-blocking connectors 
to prevent deadlocks.

Added as a separate test to confirm solution for #86.