Fix too loose check for existing data in JSON API source
`_transformCache()` incorrectly changed operation type from `'replace'` to `'add'`
if the existing data was a falsy value.

This bug can be reproduced using the following scenario:

Suppose we have a memory source and a JSON API source, which are connected to
each other using `TransformConnector`s. If we update an attribute value from
`false` to `null` in a record using the memory source, Orbit.js starts making
`PUT` requests to update the  attribute value to `null` infinitely.