Move inverse link maintenance to Cache.
Inverse link maintenance was previously performed by the MemorySource.
Moving this maintenance to the Cache allows other sources to utilize
this same logic.

This commit introduces a new `maintainInverseLinks` option on the
Cache to control whether inverse link maintenance is performed by
a particular cache.

It also removes the inconsistently applied `allowNoOps` option on
the Cache.