Cache no longer tracks changes and emits events.
The `Cache#trackChanges` option has been removed.

Cache is longer evented and no longer emits `didTransform` events.
Transform notifications are now the responsibility of the cache's
owner (i.e. the caller of `Cache#transform`).

As a result, any ancillary changes to the cache (such as reverse
and inverse link maintenance) will not bubble out of the cache.