Refactor Transformable.
This represents a significant change to the `Transformable` interface and
expectations for sources that implement it.

The `_transform` method now accepts an array of operations as an argument
instead of a single operation. Performing one operation at a time maximum was
a severe limitation for some sources, such as the `JSONAPISource`. Each source
now has the ability to merge and coalesce operations as it sees fit.

The `_transform` method should now return either a `TransformResult` or a
promise that resolves to a `TransformResult`. Previously, sources were expected
to call `didTransform` as they performed transformations, but this has been
deprecated.

The `transformed` method has been introduced for sources to call when they
experience transformations besides those requested via `_transform`. This is
useful for any sources that spontaneously receive data outside of transform
requests. The `transformed` method takes a single argument: a `TransformResult`.

The `didTransform` event is still triggered when a source changes, either from
a call to `transform` or `transformed`. Event handlers can expect a single
argument: a `Transformation`. This now public class includes the `operations`
applied to a source as well as the `result` (a `TransformResult` instance).

Furthermore, `Transformation` now contains ancestry information, which was
previously kept on each `Operation`. This allows a `TransformConnector` to
`spawn` one `Transformation` from another and thereby mark them as related.
This kind of ancestry tracking can be completely removed from sources.