Extract `query` from `find`.
OC sources should now support `query` as a separate method from `find`.
This simplifies the logic required to determine whether a request is
a query or a request for a specific individual record. It also allows
for separate requestable hooks for `find` vs. `query`.

Also, this commit explicitly removes support for `options` passed into
MemorySource and JSONAPISource requestable methods. Options, such as
`include`, were not tested in these sources and the implementations
needs re-evaluation.