Refactor Orbit.Transaction into OC.Transaction.
Instead of a `Transaction` simply observing a source, `Transaction` becomes a
source itself.

`Transaction` is a form of `MemorySource` that isolates changes from a base
source.

A transaction has its own internal cache. By default, this cache will
`retrieve` data from the base source if it's not defined. Alternatively, a
transaction can be completely isolated from the base source by setting the
`isolated` option to `true`. An isolated transaction's cache must be
manually managed.

Transactions are "begun" immediately by default. This means that they
immediately begin tracking transforms applied to them. When `commit` is
invoked, the transforms applied to the transaction will be coalesced and
applied to the base source.