Updatable#update now internally calls transform.
Instead of relying on a separate method to update the contents of a
source, `update` now internally calls `transform`.

This eliminates the need for updatable sources to declare `_update`.
Instead they should only require `_transform`.

In addition, the `update` method’s returned promise resolves to only
the single requested transform (which may be coerced from args).
It was not a clean abstraction to attempt to consolidate all the
transforms resulting from a call to `update` (via `beforeUpdate`
handlers, for instance).