Updatable#update now returns transforms applied to source.
The return value of `update` is now consistent with `fetch` and 
`transform` - in all cases a promise is returned that resolves to the 
transforms applied to a source as a direct result of the method call.
Note that this will not include transforms applied as side effects from
event handlers, such as `beforeUpdate` or `update`.