Refactor `Fetchable` -> `Pullable`.
This is part of the semantic refactor from the terms `update` + `fetch`
to the simpler terms `push` + `pull`. These names better indicate the
symmetry of these methods, which are used as part of the request flow
to either push or pull transforms.

Also, refactors `Source#transformed` -> `Source#_transformed` to clarify
that it is a private method (used “intimately” by source interfaces).