Introduce `Pushable` interface.
The `Pushable` interface adds a single method to a Source: `push`.
`push` accepts a transform as an argument and returns an array of
transforms that are applied as a result.

This is part of the semantic refactor from the terms `update` + `fetch`
to the simpler terms `push` + `pull`. These names better indicate the
symmetry of these methods, which are used as part of the request flow
to either push or pull transforms.