Enable serialization of Actions.
Actions will just track a `method` to be called with `data` as an arg,
along with a `meta` object for details such as `id`, `label`, etc.
All of these properties are serializable, unlike the `process` function
which was used formerly.

ActionQueue and Action will also track `target` - the target object for
actions. The `target` will typically be the Source (or other object)
that owns the queue.