Support timeouts for fetch.
Introduce a new option for JSONAPISource called `defaultFetchTimeout`,
after which fetch responses will be ignored and a NetworkError will be
raised.

Furthermore, the timeout can be specified as an option per request.

Note that, since `fetch` itself does not support a timeout param, this
functionality is achieved using a timeout wrapper promise. This is not
a connection timeout, but rather a response timeout.