Query expression + builder refactor
Converts query expressions to a less free form AST, to allow for
simpler and clearer implementations.

A QueryBuilder class (instead of POJO) has been created. The
builder can be used on its own or by passing a function with the builder
as an argument (e.g. `source.query(q => q.findRecords(‘planet’))`