Perform LogTruncationStrategy when queues are emptied
Instead of performing the LogTruncationStrategy on every `transform` event, observe the `complete` event for each source’s `requestQueue` and `syncQueue`.

Only perform log truncation when all queues are empty.

This prevents premature log truncation which might result in errors such as the same transform being applied twice to a source.