[jsonapi] Consider application/json to be valid content
In an attempt to be more forgiving of server responses, the JSONAPISource will attempt to parse `Content-Type: application/json` as if it is compliant JSON:API. This eliminates a common error developers hit when using the JSONAPISource.

In order to be more strict and revert to the previous behavior, override the `responseHasContent` method.

Closes #487