Introduce new `hints` object for requests
This change applies to every interface that participates in the request flow, including: `Pullable`, `Pushable`, `Queryable`, and `Updatable`.

A new `hints` argument, initialized as an empty object, is now passed to every `before` listener as a second argument. Listeners can provide “hints” by adding members to this object that will enable the source to fulfill the request.

This concept has been inspired by the need for stores to respect the sorting and filtering options of remote sources when fulfilling a query. This data needs to be passed somehow from listener to source prior to the fulfillment of the query.

Hints can also be used broadly to inform how sources process requests.