Introduce `passHints` option for `RequestStrategy`s
The `passHints` option determines whether results returned from calling `action` on the `target` source should be passed as hint data back to the `source`.

This can allow hints to inform the processing of subsequent actions on the source. For instance, a `beforeQuery` event might invoke `query` on a target, and those results could inform how the originating source performs `_query`. This might allow a target source's sorting and filtering of results to affect how the originating source processes the query.

This setting is only effective for `blocking` strategies, since only in those scenarios is processing delayed.