The cache intergrity processor was returning duplicate replateReplaceRecord/removeFromRelatedRecord operations (already hadnled by the cache consistency processor) after a removeRecord.
For a removeRecord with dependent relationships, the cache consistency processor would correctly identity the removal of the dependent records. However, the integrity processor was returning the replace/remove related record which was keeping a shell record of the removed records around making it appear as if the record was not deleted.

Update the cache integrity processor to no longer return operations after a removeRecord.