Throw when invalid filter expression.
Previously, the QueryBuilder assumed that a filter specifier without a `kind` was an attribute filter. This was true even if no attribute property was present, which could never work. This silently failed to filter anything. Furthermore, an end-developer used to other filter APIs might attempt to pass something like `{ firstName: 'Luke' }` to filter, which would fall into this silent failure path.

This commit ensures that an attribute is specified before assuming the filter is an attribute filter. In the example above, that means the well-meaning and otherwise intelligent and good-looking end developer will receive a helpful error instead of, say, shipping a non-functional filter to production.

Co-authored-by: Matt McManus <matt@mcmanus.io>