Standardize update responses
A singular result should be returned when `update` is called with a `Transform` that contains only one operation. Otherwise, an array of results should be returned.

For standard operations, this should not be a breaking change. However, if you've been defining your own custom operations that returned arrays together with hints in v0.16 beta, you'll want to test this change thoroughly.