Deprecate `_transformed` in favor of `transformed`
Now that source implementations are more likely to be responsible for calling this method to let others receive `transform` events, the time is probably right to make it more ergonomic. It is a protected method in TS, which accomplishes the same goal of signaling to developers that it should only be invoked internally within sources.