[jsonapi] Support `url` request option
Allow the `url` of a request to be overridden for all queries and transforms.

Any `url` that's passed will form a base URL that may be further amended with query params if additional options are passed that would normally include query params. Note that the `url` itself may also include query params.