Introduce UninitializedRecord interface, with optional `id`
The new `UninitializedRecord` interface can be used when adding records with the `TransformBuilder`, which will transform them to full-fledged `Record`s via the `RecordInitializer`.

`RecordInitializer#initializeRecord` has been updated to be expected to return a `Record`, instead of just initializing the passed object. The old behavior is still supported but now deprecated.

These changes should clarify typings and eliminate the need to coerce new records (`as Record`) when adding them without an assigned `id`.