Fix IndexedDBCache getInverseRelationshipsAsync
This fixes an issue with `IndexedDBCache#getInverseRelationshipsAsync` in which
the wrong index was being used to lookup items in the object store that
maintains inverse relationships (`__inverseRels__`).

This bug could result in the inability of the `AsyncCacheIntegrityProcessor` to
properly remove relationships when records themselves are removed.

The correct index, `relatedIdentity`, also needs to be added. This will need to be
handled for pre-existing databases with an IndexedDB migration.