Allow for explicit "singular" requests (queries / transforms)
`Query#expressions` now accepts either a single expression or an array of expressions.

Likewise, `Transform#operations` now accepts either a single operation or an array of operations.

This allows more explicit control over the form in which results will be returned. Data will always be returned in an array when an array is used to express expressions / operations.