MemorySource#merge - merge cache update operations if they are being tracked
If a cache is tracking update operations, which is now the default for forked caches, then when the associated source is merged, the cache's operations will be applied in a single transform.

The benefit here is that it allows caches to be manipulated synchronously, and to have all those updates applied during a merge.

If you want to continue to merge only transforms applied to a source, and not its cache, then create the fork with `source.fork({ cacheSettings: { trackUpdateOperations: false } });`.